update link, fix css, create login class and modify AbstracController to AbstractTwigController, clean code
