+++ public/assets/css/General/loginModal.css
@@ -1,7 +1,7 @@
 .modal {
     display: none;
     position: fixed;
-    z-index: 1;
+    z-index: 3;
     padding-top: 80px;
     top: 80px;
     width: 100%;
