+++ public/assets/css/setting/update_profile.css
@@ -2,6 +2,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    width: 100%;
 }
 
 .update-profile-form {
