+++ src/Controller/AbstractTwigController.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Controller;
+
+use App\Model\UserManager;
+use Twig\Environment;
+use Twig\Extension\DebugExtension;
+use Twig\Loader\FilesystemLoader;
+
+abstract class AbstractTwigController
+{
+    protected Environment $twig;
+    protected array|false $user;
+
+    /**
+     *  Initializes this class.
+     */
+    public function __construct()
+    {
+        $loader = new FilesystemLoader(APP_VIEW_PATH);
+        $this->twig = new Environment(
+            $loader,
+            [
+                'cache' => false,
+                'debug' => true,
+            ]
+        );
+
+        $this->twig->addExtension(new DebugExtension());
+
+        $userManager = new UserManager();
+        $this->user = isset($_SESSION['ID_user']) ? $userManager->getOneUser($_SESSION['ID_user']) : false;
+        $this->twig->addGlobal('user', $this->user);
+    }
+}
