+++ src/View/Navbars/SideNavbar.html.twig
@@ -15,7 +15,7 @@
 
 		<li>
 			<div class="link-animation-container">
-				<a href="" class="link-animation" title="Home" alt="Home"><img class="home-icon side-navbar-icons" src="/assets/images/Navbars/icon-home.png"><p class="hide-labels">Home</p>
+				<a href="/" class="link-animation" title="Home" alt="Home"><img class="home-icon side-navbar-icons" src="/assets/images/Navbars/icon-home.png"><p class="hide-labels">Home</p>
 				</a>
 			</div>
 		</li>
@@ -59,7 +59,7 @@
 				</li>
 				<li>
 					<div class="link-animation-container">
-						<a href="#" title="Add music" alt="Add music">
+						<a href="/setting/add_music" title="Add music" alt="Add music">
 							<p class="hide-labels link-animation">Add music</p>
 						</a>
 					</div>
