+++ src/View/Setting/Admin/admin.html.twig
@@ -1,11 +1,12 @@
 {% extends 'layout.html.twig' %}
 
+{% block title %} Wildify | Admin USERS {% endblock %}
 
 
 {% block content %}
     <section class="profile-section">
 
-        <h1 class="setting-title">>Wildify | Admin USER Page <</h1>
+        <h1 class="setting-title">> Manage USER <</h1>
 
     <section>
         {% for user in users %}
